feat: add pubg
