+++ internal/bot/listener/cmd-purge-pubg.go
@@ -1,0 +1,23 @@
+package listener
+
+import (
+	"github.com/haski007/insta-bot/pkg/emoji"
+	"github.com/sirupsen/logrus"
+
+	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
+)
+
+func (rcv *InstaBotService) cmdPurgePUBGPlayersHandler(update tgbotapi.Update) {
+	chatID := update.Message.Chat.ID
+
+	if err := rcv.storage.DeleteChat(PUBGContext, chatID); err != nil {
+		rcv.SendError(chatID, ErrInternalServerError)
+		rcv.log.WithError(err).Error("[cmdPurgePUBGPlayersHandler] delete pubg chat with members")
+		return
+	}
+
+	message := "Successfully deleted stopped to listen pubg triggers "
+	if err := rcv.SendMessage(chatID, message+emoji.Basket); err != nil {
+		logrus.WithError(err).Println("[cmdPurgePUBGPlayersHandler] send message to chat")
+	}
+}
