+++ internal/bot/listener/errors.go
@@ -7,4 +7,7 @@ const (
 	ErrNoPlayers = "There are no players registered in this chat!" +
 		"\nTo register enter /reg_csgo_players {username_1} {username_2} ..."
 	ErrNoArguments = "This command requires one or more arguments"
+
+	ErrNoPUBGPlayers = "There are no PUBG players registered in this chat!" +
+		"\nTo register enter /reg_pubg_players {username_1} {username_2} ..."
 )
